requirements.txt
absl-py==1.4.0
chex==0.0.4
contourpy==1.0.7
cycler==0.11.0
dm-haiku==0.0.4
dm-tree==0.1.8
flatbuffers==2.0.7
fonttools==4.39.3
importlib-resources==5.12.0
jax==0.2.13
jaxlib==0.1.67
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.1
numpy==1.22.4
opt-einsum==3.3.0
optax==0.0.8
packaging==23.0
pandas==2.1.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
toolz==0.12.0
typing_extensions==4.5.0
tzdata==2023.3
xmltodict==0.13.0
zipp==3.15.0
